Converted so the test system is a library with an example project. Fixed projects to use .build for intermediate and bin for output

+++ CodeCompareLib/TestSuite.cpp
@@ -36,7 +36,7 @@ unique_ptr<TestResults> TestSuite::RunTests() const
 			if (itr != PassWeights.end())
 				weight = itr->second;
 
-			for (int i = 0; i < numParams; ++i)
+			for (size_t i = 0; i < numParams; ++i)
 			{
 				auto const& parameter = Parameters[i];
 
@@ -73,7 +73,7 @@ unique_ptr<TestResults> TestSuite::RunTests() const
 			cout << endl;
 		}
 
-		for (int i = 0; i < numParams; ++i)
+		for (size_t i = 0; i < numParams; ++i)
 		{
 			summary[i].Performance /= test->GetPasses().size();
 			summary[i].MemoryUsage /= test->GetPasses().size();
